+++ src/leiningen/new/reagent.clj
@@ -8,7 +8,7 @@
 (defn reagent-dependencies [{:keys [features]}]
   (into
     [['reagent "1.1.0"]
-     ['cljs-ajax "0.8.3"]]
+     ['cljs-ajax "0.8.4"]]
     (when (some #{"+figwheel"} features)
       [['cljsjs/react "17.0.2-0"]
        ['cljsjs/react-dom "17.0.2-0"]])))